store changes integrated for named modules
